+++ src/socean.ts
@@ -7,10 +7,13 @@
  */
 import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import {
+  Commitment,
   ConfirmOptions,
   Connection,
+  Context,
   Keypair,
   PublicKey,
+  Signer,
   StakeProgram,
   Transaction,
 } from "@solana/web3.js";
@@ -73,7 +76,8 @@ export class Socean {
 
   /**
    * Signs, sends and confirms the transactions required to deposit SOL
-   * into the Socean stake pool
+   * into the Socean stake pool.
+   * Creates the scnSOL associated token account for the wallet if it doesnt exist.
    * @param walletAdapter SOL wallet to deposit SOL from
    * @param amountLamports amount to deposit in lamports
    * @param referrerPoolTokenAccount PublicKey of a scnSOL token account of the referrer for this deposit
@@ -103,6 +107,7 @@ export class Socean {
    * Creates a `TransactionSequence` that deposits SOL into Socean stake pool
    * Each inner `TransactionWithSigners` array must be executed and confirmed
    * before going to the next one.
+   * Creates the scnSOL associated token account for the wallet if it doesnt exist.
    * This is a lower-level API for compatibility, recommend using `depositSol()` instead if possible.
    * @param walletPubkey SOL wallet to deposit SOL from
    * @param amountLamports amount to deposit in lamports
@@ -173,11 +178,13 @@ export class Socean {
   /**
    * Signs, sends and confirms the transactions required to deposit a stake account
    * into the Socean stake pool
+   * Creates the scnSOL associated token account for the wallet if it doesnt exist.
    * @param walletAdapter SOL wallet to deposit SOL from
    * @param stakeAccount The stake account to deposit.
    *                     Must be active and delegated to a validator in the stake pool.
-   *                     The entire stake account is deposited. To deposit only a portion of it, prefix this instruction
-   *                     with a stake split instruction and deposit the stake account that's split off instead.
+   * @param amountLamports The amount of stake to split from `stakeAccount` to deposit.
+   *                       If not provided or 0, the entire stake account is deposited.
+   *                       Otherwise, a stake account containing `amountLamports` is first split from `stakeAccount` and then deposited.
    * @param referrerPoolTokenAccount PublicKey of a scnSOL token account of the referrer for this deposit
    * @param confirmOptions transaction confirm options for each transaction
    * @returns the transaction signatures of the transactions sent and confirmed
@@ -189,6 +196,7 @@ export class Socean {
   async depositStake(
     walletAdapter: WalletAdapter,
     stakeAccount: PublicKey,
+    amountLamports?: Numberu64,
     referrerPoolTokenAccount?: PublicKey,
     confirmOptions: ConfirmOptions = TRANSACTION_SEQUENCE_DEFAULT_CONFIRM_OPTIONS,
   ): Promise<TransactionSequenceSignatures> {
@@ -197,6 +205,7 @@ export class Socean {
     const txSeq = await this.depositStakeTransactions(
       walletPubkey,
       stakeAccount,
+      amountLamports,
       referrerPoolTokenAccount,
     );
     return this.signAndSend(walletAdapter, txSeq, confirmOptions);
@@ -207,12 +216,14 @@ export class Socean {
    * with one of the Socean stake pool's validators into Socean stake pool
    * Each inner `TransactionWithSigners` array must be executed and confirmed
    * before going to the next one.
+   * Creates the scnSOL associated token account for the wallet if it doesnt exist.
    * This is a lower-level API for compatibility, recommend using `depositStake()` instead if possible.
    * @param walletPubkey SOL wallet to deposit SOL from
    * @param stakeAccount The stake account to deposit.
    *                     Must be active and delegated to a validator in the stake pool.
-   *                     The entire stake account is deposited. To deposit only a portion of it, prefix this instruction
-   *                     with a stake split instruction and deposit the stake account that's split off instead.
+   * @param amountLamports The amount of stake to split from `stakeAccount` to deposit.
+   *                       If not provided or 0, the entire stake account is deposited.
+   *                       Otherwise, a stake account containing `amountLamports` is first split from `stakeAccount` and then deposited.
    * @param referrerPoolTokenAccount PublicKey of a scnSOL token account of the referrer for this deposit
    * @returns the deposit transaction sequence
    * @throws RpcError
@@ -222,6 +233,7 @@ export class Socean {
   async depositStakeTransactions(
     walletPubkey: PublicKey,
     stakeAccount: PublicKey,
+    amountLamports?: Numberu64,
     referrerPoolTokenAccount?: PublicKey,
   ): Promise<TransactionSequence> {
     const { value } = await this.config.connection.getParsedAccountInfo(
@@ -237,7 +249,10 @@ export class Socean {
     )
       throw new StakeAccountToDepositInvalidError("not a stake account");
     const stakeAccountInfo: ParsedStakeAccount = value.data.parsed;
-    if (!stakeAccountInfo.info.stake.delegation.voter)
+    if (
+      !stakeAccountInfo.info.stake.delegation.voter ||
+      !stakeAccountInfo.info.stake.delegation.stake
+    )
       throw new StakeAccountToDepositInvalidError(
         "stake account not delegated",
       );
@@ -269,13 +284,34 @@ export class Socean {
       this.config.stakePoolAccountPubkey,
     );
 
+    const signers: Signer[] = [];
+    let stakeAccountToDeposit = stakeAccount;
+
+    if (
+      amountLamports &&
+      !amountLamports.isZero() &&
+      amountLamports.lt(stakeAccountInfo.info.stake.delegation.stake)
+    ) {
+      const splitStakeAccount = Keypair.generate();
+      signers.push(splitStakeAccount);
+      stakeAccountToDeposit = splitStakeAccount.publicKey;
+
+      const splitTx = StakeProgram.split({
+        stakePubkey: stakeAccount,
+        authorizedPubkey: walletPubkey,
+        splitStakePubkey: splitStakeAccount.publicKey,
+        lamports: amountLamports.toNumber(),
+      });
+      tx.add(splitTx);
+    }
+
     const ix = depositStakeInstruction(
       this.config.stakePoolProgramId,
       this.config.stakePoolAccountPubkey,
       stakePool.account.data.validatorList,
       stakePool.account.data.depositAuthority,
       stakePoolWithdrawAuthority,
-      stakeAccount,
+      stakeAccountToDeposit,
       vsa,
       stakePool.account.data.reserveStake,
       poolTokenTo,
@@ -289,7 +325,7 @@ export class Socean {
       [
         {
           tx,
-          signers: [],
+          signers,
         },
       ],
     ];
@@ -591,6 +627,66 @@ export class Socean {
       main,
       transient,
     };
+  }
+
+  /**
+   * Register an onAccountChange websocket listener for the stake pool account
+   * and returns the websocket clientSubscriptionId
+   * To remove, call this.config.connection.removeAccountChangeListener(clientSubscriptionId)
+   * @param callback
+   * @param commitment
+   * @returns clientSubscriptionId
+   */
+  onStakePoolChange(
+    callback: (stakePoolAccount: StakePoolAccount, context: Context) => void,
+    commitment?: Commitment,
+  ): number {
+    return this.config.connection.onAccountChange(
+      this.config.stakePoolAccountPubkey,
+      (account, context) => {
+        const stakePoolAccount = getStakePoolFromAccountInfo(
+          this.config.stakePoolAccountPubkey,
+          account,
+        );
+        callback(stakePoolAccount, context);
+      },
+      commitment,
+    );
+  }
+
+  /**
+   * Register an onAccountChange websocket listener for the validator list account
+   * and returns the websocket clientSubscriptionId
+   * To remove, call this.config.connection.removeAccountChangeListener(clientSubscriptionId)
+   * @param callback
+   * @param commitment
+   * @returns clientSubscriptionId
+   * @throws RpcError
+   * @throws AccountDoesNotExistError if stake pool does not exist
+   */
+  async onValidatorListChange(
+    callback: (
+      validatorListAccount: ValidatorListAccount,
+      context: Context,
+    ) => void,
+    commitment?: Commitment,
+  ): Promise<number> {
+    const {
+      account: {
+        data: { validatorList },
+      },
+    } = await this.getStakePoolAccount();
+    return this.config.connection.onAccountChange(
+      validatorList,
+      (account, context) => {
+        const validatorListAccount = getValidatorListFromAccountInfo(
+          validatorList,
+          account,
+        );
+        callback(validatorListAccount, context);
+      },
+      commitment,
+    );
   }
 }