+++ CHANGELOG.md
@@ -6,7 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Changed
+- `depositStake` and `depositStakeTransactions` now take an additional `amount` argument that can be used to split an active stake account before depositing. 
+
 ### Added
+- `onStakePoolChange` and `onValidatorListChange`
 - Export account schema types and classes
 
 ## [0.3.0] - 2022-04-28